test(block): separate tests for get_block_children function
