+++ tests/block/mod.rs
@@ -1,2 +1,2 @@
 mod crud_audio_block;
-mod get_block;
+mod get_block_children;
